handle the hissa for large digits
handle the hissa for large digits in calculateNetPart funciton

+++ SDC_Application/AL/frmFardeBadar.cs
@@ -974,9 +974,9 @@ namespace SDC_Application.AL
 
         #region Save malik entry
 
-        private float calculateNetPart(string valueInBata)
+        private double calculateNetPart(string valueInBata)
         {
-            float NetPart = 0;
+            double NetPart = 0;
             try
             {
                 if (valueInBata.Contains('/'))
@@ -1002,7 +1002,7 @@ namespace SDC_Application.AL
                 }
                 else
                 {
-                    NetPart = float.Parse(valueInBata);
+                    NetPart = double.Parse(valueInBata);
                 }
             }
             catch (Exception ex)
